renewed customerpo,supplierpo,supplierquotation,materialinward, materialoutward

+++ src/main/java/com/accurate/erp/model/inventory/MaterialOutwardDO.java
@@ -134,25 +134,25 @@ public class MaterialOutwardDO {
 	String month;
 	
 	@Column(name="transportGst")
-	Integer transportGst;
+	BigDecimal transportGst;
 	
 	
 	 @Column(name="otherchargesGst") 
-	 Integer additionalChargesGst;
+	 BigDecimal additionalChargesGst;
 	
 	
-	public Integer getTransportGst() {
+	public BigDecimal getTransportGst() {
 		return transportGst;
 	}
 
-	public void setTransportGst(Integer transportGst) {
+	public void setTransportGst(BigDecimal transportGst) {
 		this.transportGst = transportGst;
 	}
 
 
-	  public Integer getAdditionalChargesGst() { return additionalChargesGst; }
+	  public BigDecimal getAdditionalChargesGst() { return additionalChargesGst; }
 	  
-	  public void setAdditionalChargesGst(Integer additionalChargesGst) {
+	  public void setAdditionalChargesGst(BigDecimal additionalChargesGst) {
 	 this.additionalChargesGst = additionalChargesGst; }
